rename initCacheStorage to cache.ts
