+++ core/modules/cart/test/unit/index.spec.ts
@@ -4,7 +4,7 @@ jest.mock('../../store', () => ({}));
 jest.mock('@vue-storefront/module', () => ({ createModule: jest.fn(() => ({ module: 'cart' })) }));
 jest.mock('../../helpers/cartCacheHandler', () => ({ cartCacheHandlerFactory: jest.fn() }))
 jest.mock('@vue-storefront/core/helpers', () => ({ isServer: false }))
-jest.mock('@vue-storefront/core/helpers/initCacheStorage', () => ({ initCacheStorage: jest.fn() }));
+jest.mock('@vue-storefront/core/helpers/cache', () => ({ initCacheStorage: jest.fn() }));
 
 describe('Cart Module', () => {
   it('can be initialized', () => {
