+++ core/modules/offline-order/components/CancelOrders.ts
@@ -1,10 +1,10 @@
 import { Logger } from '@vue-storefront/core/lib/logger'
-import { initCacheStorage } from '@vue-storefront/core/helpers/initCacheStorage'
+import { StorageManager } from '@vue-storefront/core/store/lib/storage-manager'
 
 export const CancelOrders = {
   methods: {
     cancelOrders () {
-      const ordersCollection = initCacheStorage('orders', false, true)
+      const ordersCollection = StorageManager.get('orders')
       ordersCollection.iterate((order, id, iterationNumber) => {
         if (!order.transmited) {
           ordersCollection.removeItem(id)
